+++ app/src/main/java/com/example/recipeapp/Adapters/IngredientsAdapter.java
@@ -82,7 +82,6 @@ public class IngredientsAdapter extends RecyclerView.Adapter<IngredientsAdapter.
                     Log.i(TAG, "Couldn't retrieve substitutes");
                 }
             });
-
         }
     }
 }
