+++ app/src/main/java/com/example/recipeapp/IngredientsActivity.java
@@ -34,6 +34,9 @@ public class IngredientsActivity extends AppCompatActivity {
 
         ingredients = Parcels.unwrap(getIntent().getParcelableExtra(DetailsActivity.INGREDIENTS));
 
+        // Since there are limited number of call on free api, and a lot of ingredients, this line is used for testing.
+        //ingredients = ingredients.stream().limit(2).collect(Collectors.toList());
+
         adapter = new IngredientsAdapter(ingredients);
 
         binding.rvIngredients.setLayoutManager(new LinearLayoutManager(this));
