+++ app/src/main/java/com/example/recipeapp/Utils/Recommendation.java
@@ -20,6 +20,7 @@ import com.parse.ParseException;
 import com.parse.ParseObject;
 import com.parse.ParseUser;
 
+import java.io.IOException;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
@@ -50,6 +51,8 @@ public class Recommendation {
         }
 
         RecipeApi service = RetrofitClientInstance.getRetrofitInstance().create(RecipeApi.class);
+        // I think it's ok to live a call here, since Recommendations acts as a repository
+        // Should I rename it?
         Call<Envelope<List<Recipe>>> call = service.getSortedRecipes(BuildConfig.API_KEY,
                 getListRecommendation(currentPreferences, generalPreferences, Preferences.KEY_USER_CUISINE),
                 getListRecommendation(currentPreferences, generalPreferences, Preferences.KEY_USER_DIET),
@@ -68,7 +71,7 @@ public class Recommendation {
                 List<Recipe> recipes = response.body().results;
                 Executor.Builder builder = new Executor.Builder();
                 for(Recipe recipe : recipes) {
-                    builder = builder.add(() -> recipe.getTaste(current));
+                    builder = builder.add(() -> getTaste(recipe, current));
                 }
                 builder = builder.callback(() -> {
                     Collections.sort(recipes, Comparator.comparingDouble(Recipe::getUserRating));
@@ -82,6 +85,17 @@ public class Recommendation {
                 Log.e(TAG, "Failed " + t);
             }
         };
+    }
+
+    public static void getTaste(Recipe recipe, Preferences current) {
+        RecipeApi service = RetrofitClientInstance.getRetrofitInstance().create(RecipeApi.class);
+        Call<Taste> call = service.getTasteById(recipe.getId(), BuildConfig.API_KEY);
+        try {
+            Response<Taste> response = call.execute();
+            recipe.setUserRating(Recommendation.getRecipeDistance(recipe, response.body(), current));
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
     }
 
     public static String getListRecommendation(Preferences current, Preferences general, String key) {
